root reducer moved to separate file
